[refactor] fetch label 구조 변경
